+++ clients/admin-ui/src/features/plus/plus.slice.ts
@@ -250,6 +250,13 @@ const plusApi = baseApi.injectEndpoints({
       }),
       providesTags: ["Dictionary"],
     }),
+    getSystemHistory: build.query<
+      SystemHistoryResponse,
+      { system_key: string }
+    >({
+      query: (params) => ({ url: `plus/system/${params.system_key}/history` }),
+      providesTags: () => ["System History"],
+    }),
   }),
 });
 
@@ -275,6 +282,7 @@ export const {
   useGetAllCustomFieldDefinitionsQuery,
   useGetAllowListQuery,
   useGetAllDictionaryEntriesQuery,
+  useGetSystemHistoryQuery,
 } = plusApi;
 
 export const selectHealth: (state: RootState) => HealthCheck | undefined =
