+++ clients/admin-ui/src/types/api/models/SystemHistory.ts
@@ -1,0 +1,7 @@
+export type SystemHistory = {
+  edited_by: string;
+  system_id: string;
+  before: object;
+  after: object;
+  created_at: string;
+};
