+++ clients/admin-ui/src/types/api/models/SystemHistoryResponse.ts
@@ -1,0 +1,8 @@
+import { SystemHistory } from "./SystemHistory";
+
+export type SystemHistoryResponse = {
+  items: SystemHistory[];
+  total: number;
+  page: number;
+  size: number;
+};
